refactor(publisher): hoist the htmlAttributes emit into core to break the module cycle
renderLoop importing @modules/base/shared/htmlAttributes closed a
publisher->modules->publisher cycle through the modules escape shim. The
publish-side string emit (htmlAttributesAttr) now lives in @core/publisher,
the pure normalisation (normalizeHtmlAttributes / htmlAttributesForReact)
in @core/htmlAttributes next to the sanitise gate it wraps, and the modules
shared file keeps only the module-engine control and schema options.
Also registers htmlAttributes in the base.loop defaults contract test.
