+++ src/__tests__/base-modules.test.ts
@@ -172,6 +172,7 @@ describe('base.loop — module contract specifics', () => {
       'customTag',
       'direction',
       'filters',
+      'htmlAttributes',
       'limit',
       'offset',
       'orderBy',
