+++ src/core/htmlAttributes/index.ts
@@ -1,7 +1,9 @@
 export {
+  htmlAttributesForReact,
   isEventHandlerAttributeName,
   isRenderableHtmlAttributeName,
   isReservedRuntimeDataAttributeName,
   normalizeHtmlAttributeName,
+  normalizeHtmlAttributes,
   sanitizeRenderableHtmlAttribute,
 } from './attributes'
