+++ src/modules/base/container/ContainerEditor.tsx
@@ -34,7 +34,7 @@
 import React from 'react'
 import type { ModuleComponentProps } from '@core/module-engine'
 import { resolveHtmlTag, VOID_HTML_ELEMENTS } from '@modules/base/utils/htmlTag'
-import { htmlAttributesForReact } from '@modules/base/shared/htmlAttributes'
+import { htmlAttributesForReact } from '@core/htmlAttributes'
 import { CanvasModulePlaceholder } from '@ui/components/CanvasModulePlaceholder'
 import { ContainerSolidIcon } from 'pixel-art-icons/icons/container-solid'
 import type { ContainerStoredProps } from './props'
