+++ src/modules/base/container/index.ts
@@ -18,9 +18,9 @@ import {
   VOID_HTML_ELEMENTS,
 } from '@modules/base/utils/htmlTag'
 import {
-  htmlAttributesAttr,
   htmlAttributesControl,
 } from '@modules/base/shared/htmlAttributes'
+import { htmlAttributesAttr } from '@core/publisher'
 import { Value } from '@core/utils/typeboxHelpers'
 import { ContainerEditor } from './ContainerEditor'
 import { ContainerPropsSchema, type ContainerStoredProps } from './props'
