+++ src/modules/base/link/LinkEditor.tsx
@@ -10,7 +10,7 @@
 import React from 'react'
 import type { ModuleComponentProps } from '@core/module-engine'
 import { anchorRel } from '@modules/base/shared/anchorTarget'
-import { htmlAttributesForReact } from '@modules/base/shared/htmlAttributes'
+import { htmlAttributesForReact } from '@core/htmlAttributes'
 import { inlineEditableElementProps } from '@modules/base/shared/inlineText'
 import { linkUsesChildren } from './content'
 import type { LinkStoredProps } from './props'
