+++ src/modules/base/text/index.ts
@@ -9,9 +9,9 @@ import type { ModuleDefinition } from '@core/module-engine'
 import { TextStartTIcon } from 'pixel-art-icons/icons/text-start-t'
 import { Value } from '@core/utils/typeboxHelpers'
 import {
-  htmlAttributesAttr,
   htmlAttributesControl,
 } from '@modules/base/shared/htmlAttributes'
+import { htmlAttributesAttr } from '@core/publisher'
 import { textToBreakHtml } from '@modules/base/shared/inlineText'
 import { TextEditor } from './TextEditor'
 import { normalizeTag } from './tags'
